requirements.txt
mcp>=1.0.0
openai>=1.0.0
langchain>=0.1.0
langchain-openai>=0.1.0
langchain-community>=0.1.0
langchain-core>=0.1.0
langchain-text-splitters>=0.1.0
langchain-chroma>=0.1.0
chromadb>=0.4.0
pydantic>=2.0.0
fastapi>=0.100.0
uvicorn>=0.20.0
python-dotenv>=1.0.0
requests>=2.31.0
aiohttp>=3.8.0
tiktoken>=0.5.0
sentence-transformers>=2.2.0
numpy>=1.24.0
pandas>=2.0.0
pytest-asyncio>=0.21.0
PyPDF2>=3.0.0
